Module Tezos_client_010_PtGRANAD.Managed_contract

val get_contract_manager : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Retrieve the manager key in a contract storage. The storage has to be of type `pair key_hash 'a`.

val build_delegate_operation : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?fee:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash option -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.transaction Annotated_manager_operation.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Builds a delegation operation ready for injection

val set_delegate : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?branch:int -> fee_parameter:Injection.fee_parameter -> ?fee:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> source:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash option -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.transaction Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.manager Injection.result, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Set the delegate of a manageable contract. For a contract with a `do`entrypoint, it builds the lambda that set the provided delegate. `~source` has to be the registered manager of the contract.

val build_transaction_operation : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> destination:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> ?entrypoint:string -> ?arg:string -> amount:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> ?gas_limit:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.counter -> unit -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.transaction Annotated_manager_operation.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Builds a transaction operation ready for injection

val transfer : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?branch:int -> source:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> contract:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> destination:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> ?entrypoint:string -> ?arg:string -> amount:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> ?gas_limit:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.counter -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.transaction Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Kind.manager Injection.result * Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t list, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Perform a transfer on behalf of a managed contract . For a contract with a `do`entrypoint, it builds the lambda that does the requested operation. `~source` has to be the registered manager of the contract.

val build_lambda_for_set_delegate : delegate:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash option -> string

val build_lambda_for_transfer_to_implicit : destination:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash -> amount:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> string

val build_lambda_for_transfer_to_originated : destination:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t -> entrypoint:string -> amount:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t -> parameter_type:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Script.expr -> parameter:Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Script.expr -> string