Class Protocol_client_context.wrap_rpc_context

The class wrap_rpc_context is a wrapper class used by the proxy mode clients. From a general-purpose RPC_context.json t, the class is augmented with shell services to provide RPC calls that are protocol-dependent.

inherit Tezos_rpc.RPC_context.json

inherit (Tezos_shell_services.Shell_services.chain * Tezos_shell_services.Shell_services.block) Tezos_protocol_010_PtGRANAD.Protocol.Environment.RPC_context.simple