Module Proxy.ProtoRpc

val split_key : Proxy_context.M.key -> (Proxy_context.M.key * Proxy_context.M.key) option

When about to do a RPC request, how to modify the request; to possibly request a parent tree of the key, to batch successive requests into one.

For example when requesting baking_rights, there's the following sequence of requests (for carthage):

v1;constants v1;first_level

To speed things up, this function returns "v1" on any of these keys, to do a single request instead of two. This list is arbitrary. It was built by initial input from @klakplok and by experimenting.

This function guarantees that if it returns Some(prefix, suffix) then prefix @@ suffix = key.

val do_rpc : Tezos_proxy.Proxy.proxy_getter_input -> Proxy_context.M.key -> (Tezos_shell_services.Block_services.raw_context, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t